# openapi/blobyard-core.openapi.yaml
openapi: 3.1.0
info:
  title: Blob Yard Core API
  version: 1.0.0
  description:
    The stable API for durable object storage and scoped access. Stored files do not expire when a
    transfer grant or public capability expires.
servers:
  - url: https://api.blobyard.com/v1
security:
  - bearerAuth: []
paths:
...
  /yards:
    get:
      operationId: listWebYards
      description: Lists active Web Yards for one workspace project.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ListWebYardsSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - yard
          - list
      x-blobyard-mcp:
        tool: blobyard_list_web_yards
      parameters:
        - in: query
          name: project
          required: true
          schema:
            $ref: ./blobyard-shared.openapi.yaml#/components/schemas/Slug
        - in: query
          name: workspace
          required: true
          schema:
            $ref: ./blobyard-shared.openapi.yaml#/components/schemas/Slug
...
  /yards/deploys/finalise:
    post:
      operationId: finaliseWebYardDeploy
      description: Snapshots a ready manifest and promotes it when no newer deploy has succeeded.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/FinaliseWebYardDeploySuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - deploy
      x-blobyard-mcp:
        tool: blobyard_deploy_web_yard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./blobyard-shared.openapi.yaml#/components/schemas/FinaliseWebYardDeployRequest
  /yards/deploys/fail:
    post:
      operationId: failWebYardDeploy
      description: Marks an incomplete deploy as failed and schedules its manifest for cleanup.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/FailWebYardDeploySuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - deploy
      x-blobyard-mcp:
        tool: blobyard_deploy_web_yard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./blobyard-shared.openapi.yaml#/components/schemas/FailWebYardDeployRequest
  /yards/rollback:
    post:
      operationId: rollbackWebYard
      description:
        Promotes a retained successful deploy. Workspace administrator access is required.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/RollbackWebYardSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - yard
          - rollback
      x-blobyard-mcp:
        tool: blobyard_rollback_web_yard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./blobyard-shared.openapi.yaml#/components/schemas/RollbackWebYardRequest
  /yards/delete:
    post:
      operationId: deleteWebYard
      description:
        Deletes a Web Yard and schedules retained manifest objects for cleanup. Workspace
        administrator access is required.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/DeleteWebYardSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - yard
          - delete
      x-blobyard-mcp:
        tool: blobyard_delete_web_yard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./blobyard-shared.openapi.yaml#/components/schemas/DeleteWebYardRequest
...
components:
  securitySchemes:
    bearerAuth:
      $ref: ./blobyard-shared.openapi.yaml#/components/securitySchemes/bearerAuth
    githubOidc:
      $ref: ./blobyard-shared.openapi.yaml#/components/securitySchemes/githubOidc
...